+++ src/graphql/schema/connection/index.ts
@@ -19,27 +19,10 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { cacheSchema } from './cache'
-import { connectionSchema } from './connection'
-import { dateTimeSchema } from './date-time'
-import { instanceSchema } from './instance'
-import { licenseSchema } from './license'
-import { notificationSchema } from './notification'
-import { Schema } from './utils'
-import { uuidSchema } from './uuid'
+import { Schema } from '../utils'
+import { typeDefs } from './type-defs'
 
-export * from './connection'
-export * from './date-time'
-export * from './instance'
-export * from './license'
-export * from './uuid'
+export * from './types'
+export * from './utils'
 
-export const schema = Schema.merge(
-  connectionSchema,
-  cacheSchema,
-  dateTimeSchema,
-  instanceSchema,
-  licenseSchema,
-  notificationSchema,
-  uuidSchema
-)
+export const connectionSchema = new Schema({}, [typeDefs])