+++ src/graphql/schema/connection/types.ts
@@ -19,27 +19,28 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { cacheSchema } from './cache'
-import { connectionSchema } from './connection'
-import { dateTimeSchema } from './date-time'
-import { instanceSchema } from './instance'
-import { licenseSchema } from './license'
-import { notificationSchema } from './notification'
-import { Schema } from './utils'
-import { uuidSchema } from './uuid'
+export interface Connection<T> {
+  edges: Cursor<T>[]
+  nodes: T[]
+  totalCount: number
+  pageInfo: PageInfo
+}
 
-export * from './connection'
-export * from './date-time'
-export * from './instance'
-export * from './license'
-export * from './uuid'
+export interface ConnectionPayload {
+  after?: string
+  before?: string
+  first?: number
+  last?: number
+}
 
-export const schema = Schema.merge(
-  connectionSchema,
-  cacheSchema,
-  dateTimeSchema,
-  instanceSchema,
-  licenseSchema,
-  notificationSchema,
-  uuidSchema
-)
+export interface Cursor<T> {
+  cursor: string
+  node: T
+}
+
+export interface PageInfo {
+  hasNextPage: boolean
+  hasPreviousPage: boolean
+  startCursor: string | null
+  endCursor: string | null
+}